+++ rest_server/controllers/context/context_meta.go
@@ -2,6 +2,6 @@ package context
 
 // social 리스트 정보 요청
 type SocialInfo struct {
-	SocialType int64
+	SocialType int
 	SocialName string
 }
